refactor: move to repo utils
